+++ src/EventsManager.php
@@ -151,15 +151,11 @@ final class EventsManager
             $handler = $iterator->current();
             $iterator->next();
 
-            if (is_object($handler)) {
-                if ($handler instanceof \Closure) {
-                    $arguments = $arguments ?: [$event];
-                    $status = call_user_func_array($handler, $arguments);
-                } else {
-                    if (method_exists($handler, $eventHappening)) {
-                        $status = $handler->{$eventHappening}($event);
-                    }
-                }
+            if (is_callable($handler)) {
+                $arguments = $arguments ?: [$event];
+                $status = call_user_func_array($handler, $arguments);
+            } elseif (is_object($handler) && method_exists($handler, $eventHappening)) {
+                $status = $handler->{$eventHappening}($event);
             }
         }