+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
  - Added getting queues by priority range
  
+## [3.9.1]
+### Fixed
+-fixed stop worker if limit time is reached after running
+
+## [3.9.0]
+### Added
+- Added limits for workers by memory, execution time and tasks count.
+
+## [3.8.3]
+### Fixed
+- Fixed removing completed jobs history
+
+## [3.8.2]
+### Fixed
+- Fire event BeforeFork before fork
+- Added deregister worker for SIGQUIT
+- Fix method name beforeFirstWork
+ 
 ## [3.8.1]
 ### Added
 - Added getting workers by range.
@@ -206,7 +224,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - Initial stable release
 
-[Unreleased]: https://github.com/pdffiller/qless-php/compare/v3.8.1...HEAD
+[Unreleased]: https://github.com/pdffiller/qless-php/compare/v3.9.1...HEAD
+[3.9.0]: https://github.com/pdffiller/qless-php/compare/v3.9.0...v3.9.1
+[3.9.0]: https://github.com/pdffiller/qless-php/compare/v3.8.3...v3.9.0
+[3.8.3]: https://github.com/pdffiller/qless-php/compare/v3.8.2...v3.8.3
+[3.8.2]: https://github.com/pdffiller/qless-php/compare/v3.8.1...v3.8.2
 [3.8.1]: https://github.com/pdffiller/qless-php/compare/v3.8.0...v3.8.1
 [3.8.0]: https://github.com/pdffiller/qless-php/compare/v3.7.0...v3.8.0
 [3.7.0]: https://github.com/pdffiller/qless-php/compare/v3.6.1...v3.7.0